requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-PyMongo==2.3.0
httplib2==0.18.1
importlib-metadata==1.6.1
install==1.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.1
kazoo==2.5.0
MarkupSafe==1.1.1
more-itertools==8.4.0
oauth2client==4.1.3
packaging==20.4
pluggy==0.13.1
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pykafka==2.8.0
pymongo==3.10.1
pyparsing==2.4.7
pytest==5.4.3
rsa==4.6
six==1.15.0
tabulate==0.8.7
wcwidth==0.2.4
Werkzeug==1.0.1
zipp==3.1.0
